requirements.txt
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
codecov==2.0.15
colorama==0.3.9
coverage==4.4.2
cycler==0.10.0
decorator==4.1.2
dj-database-url==0.4.2
Django==2.0
django-bootstrap-datetimepicker==1.5
django-bootstrap3==9.1.0
django-chartjs==1.2
django-datetime-widget==0.9.3
django-excel==0.0.10
django-widget-tweaks==1.4.1
enum34==1.1.6
et-xmlfile==1.0.1
flake8==3.5.0
future==0.16.0
google-api-python-client==1.6.5
gspread==2.0.0
gunicorn==19.7.1
html5lib==0.9999999
httplib2==0.11.0
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
lml==0.0.1
lxml==4.1.1
Markdown==2.6.10
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
mpmath==1.0.0
mysqlclient==1.3.12
nltk==3.2.5
nose==1.3.7
numpy==1.13.3
oauth2client==4.1.2
odfpy==1.3.6
opencv-python==3.3.0.10
openpyxl==2.5.0
parso==0.1.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pyexcel==0.5.7
pyexcel-ezodf==0.3.4
pyexcel-handsontable==0.0.1
pyexcel-io==0.5.6
pyexcel-ods==0.5.2
pyexcel-ods3==0.5.2
pyexcel-pygal==0.0.2
pyexcel-webio==0.1.4
pyexcel-xls==0.5.5
pyexcel-xlsx==0.5.5
pyflakes==1.6.0
pygal==2.4.0
pyglet==1.3.0
Pygments==2.2.0
PyMySQL==0.8.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-gflags==3.1.2
pytz==2017.3
reportlab==3.4.0
requests==2.18.4
rsa==3.4.2
scipy==1.0.0
simplegeneric==0.8.1
sip==4.19.6
six==1.11.0
sympy==1.1.1
texttable==1.2.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.13
whitenoise==3.3.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
